+++ packages/internals/src/get-generators/generatorResolver/predefinedGeneratorResolvers.ts
@@ -5,9 +5,9 @@ import fs from 'fs'
 import hasYarn from 'has-yarn'
 import path from 'path'
 
-import { logger } from '.'
-import { getCommandWithExecutor } from './utils/getCommandWithExecutor'
-import { resolvePkg } from './utils/resolve'
+import { logger } from '../..'
+import { getCommandWithExecutor } from '../../utils/getCommandWithExecutor'
+import { resolvePkg } from '../../utils/resolve'
 
 const debug = Debug('prisma:generator')
