+++ packages/internals/src/get-generators/getGenerators.ts
@@ -13,8 +13,6 @@ import path from 'path'
 
 import { getConfig, getDMMF } from '..'
 import { Generator } from '../Generator'
-import type { GeneratorPaths } from '../predefinedGeneratorResolvers'
-import { predefinedGeneratorResolvers } from '../predefinedGeneratorResolvers'
 import { resolveOutput } from '../resolveOutput'
 import { extractPreviewFeatures } from '../utils/extractPreviewFeatures'
 import { mapPreviewFeatures } from '../utils/mapPreviewFeatures'
@@ -23,6 +21,8 @@ import { missingModelMessage, missingModelMessageMongoDB } from '../utils/missin
 import { parseBinaryTargetsEnvValue, parseEnvValue } from '../utils/parseEnvValue'
 import { pick } from '../utils/pick'
 import { printConfigWarnings } from '../utils/printConfigWarnings'
+import type { GeneratorPaths } from './generatorResolver/predefinedGeneratorResolvers'
+import { predefinedGeneratorResolvers } from './generatorResolver/predefinedGeneratorResolvers'
 import { binaryTypeToEngineType } from './utils/binaryTypeToEngineType'
 import { checkFeatureFlags } from './utils/check-feature-flags/checkFeatureFlags'
 import { getBinaryPathsByVersion } from './utils/getBinaryPathsByVersion'
